Remove assumption of aligned size with cmsghdr
Previously sending of ancillary data assumed that
CMSG_ALIGN(cp->msg_len) equals CMSG_SPACE(original_data_size). This
happened to be true since CMSG_ALIGN(sizeof(struct cmsghdr)) equals
sizeof(struct cmsghdr).

However, when the data is initially created this assumption is not used
but the correct code is used to calculate the sizes.

This commit changes the sending so that it calculates the initial
arguments used when the data is created and uses the same CMSG_SPACE to
calculate the final needed size like when the data was created.