Copy the whole ancillary data when sending
Ancillary data is stored in internal buffer as properly created cmsghdr
objects. Copy them completely to the buffer as it was initially created
instead of trying to optimize and only copy the bytes as indicated by
it's internal size argument.

This gives control to the initial creation of the ancillary data of what
exact bytes get transmitted regardless of padding or other aspects in
the creation part.

This avoids leaving uninitialized bytes to the internal buffer due to
this size difference. It doesn't change the side that
tcp_send_ancillary_data leaves these padding bytes uninitialized.