Initialize ancillary data padding with zero
Valgrind noticed that these padding bytes are sent to sendmsg syscall
uninitialized. Those bytes are actually sent to the kernel and to the
other process and it seems a bit problematic to send uninitialized bytes
to other processes.

However, this may not pose a problem as the recipient is not supposed to
read those bytes with these messages but it's still better to set them
to zero to have explicitly specified data that we send to the other
processes instead of random values in memory.